move modal sections tests to its components test file
