+++ src/hooks/__tests__/useModalContext.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import { useModalContext, ModalContextProvider } from '../useModalContext'
-import { ModalBody, ModalHeader, ModalSidebar } from '../../components/Modal'
 import { createMockModalContext } from '../../test/utils/createMockModalContext'
 
 function TestComponent() {
@@ -42,85 +41,4 @@ it('should throw an error when used outside ModalContextProvider', () => {
   expect(renderWithoutProvider).toThrow(
     'Modal subcomponents (ModalHeader, ModalBody, ModalContainer, ModalFooter, ModalCloseButton) must be used inside <Modal>'
   )
-})
-
-it('should call setSectionState when modal has header', () => {
-  const mockSetSectionState = jest.fn()
-
-  const mockContextValue = createMockModalContext({ setSectionState: mockSetSectionState })
-
-  render(
-    <ModalContextProvider value={mockContextValue}>
-      <ModalHeader title='Modal title' />
-    </ModalContextProvider>
-  )
-
-  expect(mockSetSectionState).toHaveBeenCalledTimes(1)
-  expect(mockSetSectionState).toHaveBeenCalledWith('hasHeader', true)
-})
-
-it('should call setSectionState when modal has left sidebar', () => {
-  const mockSetSectionState = jest.fn()
-
-  const mockContextValue = createMockModalContext({ setSectionState: mockSetSectionState })
-
-  render(
-    <ModalContextProvider value={mockContextValue}>
-      <ModalSidebar position='left' />
-      <ModalBody>Body content</ModalBody>
-    </ModalContextProvider>
-  )
-
-  expect(mockSetSectionState).toHaveBeenCalledTimes(1)
-  expect(mockSetSectionState).toHaveBeenCalledWith('hasLeftSidebar', true)
-})
-
-it('should call setSectionState when modal has right sidebar', () => {
-  const mockSetSectionState = jest.fn()
-
-  const mockContextValue = createMockModalContext({ setSectionState: mockSetSectionState })
-
-  render(
-    <ModalContextProvider value={mockContextValue}>
-      <ModalSidebar position='right' />
-      <ModalBody>Body content</ModalBody>
-    </ModalContextProvider>
-  )
-
-  expect(mockSetSectionState).toHaveBeenCalledTimes(1)
-  expect(mockSetSectionState).toHaveBeenCalledWith('hasRightSidebar', true)
-})
-
-it('should not call setSectionState when modal has no optional sections (header, left sidebar, right sidebar)', () => {
-  const mockSetSectionState = jest.fn()
-
-  const mockContextValue = createMockModalContext({ setSectionState: mockSetSectionState })
-
-  render(
-    <ModalContextProvider value={mockContextValue}>
-      <ModalBody>Body content</ModalBody>
-    </ModalContextProvider>
-  )
-
-  expect(mockSetSectionState).not.toHaveBeenCalled()
-})
-
-it('should call setSectionState three times when modal has header, left sidebar and right sidebar', () => {
-  const mockSetSectionState = jest.fn()
-
-  const mockContextValue = createMockModalContext({ setSectionState: mockSetSectionState })
-
-  render(
-    <ModalContextProvider value={mockContextValue}>
-      <ModalHeader title='Modal title' />
-      <ModalSidebar position='left' />
-      <ModalBody>Body content</ModalBody>
-      <ModalSidebar position='right' />
-    </ModalContextProvider>
-  )
-
-  expect(mockSetSectionState).toHaveBeenCalledTimes(3)
-  expect(mockSetSectionState).toHaveBeenCalledWith('hasHeader', true)
-  expect(mockSetSectionState).toHaveBeenCalledWith('hasLeftSidebar', true)
-  expect(mockSetSectionState).toHaveBeenCalledWith('hasRightSidebar', true)
 })
